requirements.txt
Flask~=3.1.0
rdflib~=7.1.2
groq~=0.15.0
python-dotenv~=1.0.1